\name{erbsViz-package}
\alias{erbsViz-package}
\alias{erbsViz}
\docType{package}
\title{
\packageTitle{erbsViz}
}
\description{
\packageDescription{erbsViz}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{erbsViz}
\packageIndices{erbsViz}
~~ An overview of how to use the package, including ~~
~~ the most important functions ~~
}
\author{
\packageAuthor{erbsViz}

Maintainer: \packageMaintainer{erbsViz}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
